Review updates

+++ reference/base/batch_struct.hpp
@@ -67,7 +67,7 @@ inline gko::batch_multi_vector::uniform_batch<const ValueType> get_batch_struct(
     const BatchMultiVector<ValueType>* const op)
 {
     return {op->get_const_values(), op->get_num_batch_entries(),
-            op->get_common_size()[1],
+            static_cast<int>(op->get_common_size()[1]),
             static_cast<int>(op->get_common_size()[0]),
             static_cast<int>(op->get_common_size()[1])};
 }
@@ -81,7 +81,7 @@ inline gko::batch_multi_vector::uniform_batch<ValueType> get_batch_struct(
     BatchMultiVector<ValueType>* const op)
 {
     return {op->get_values(), op->get_num_batch_entries(),
-            op->get_common_size()[1],
+            static_cast<int>(op->get_common_size()[1]),
             static_cast<int>(op->get_common_size()[0]),
             static_cast<int>(op->get_common_size()[1])};
 }
